pyproject.toml
[tool.poetry]
name = "monte-carlo-simulation"
version = "0.1.0"
description = "monte carlo simulation"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
    { include = "libs" }
]

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.10.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.6.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^25.1.0"
pre-commit = "^4.1.0"
pylint = "^3.3.4"
pydocstyle = "^6.1.1"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
